requirements.txt
wordcloud
nltk
pillow
numpy
beautifulsoup4